+++ src/main/java/wooteco/subway/auth/exception/AuthorizationException.java
@@ -1,9 +1,5 @@
-package wooteco.subway.auth.application;
+package wooteco.subway.auth.exception;
 
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
 public class AuthorizationException extends RuntimeException {
     public AuthorizationException() {
     }
